Simple var initialization and expression framework ready.
The interpreter now understands `var x = <...>`. It doesn't yet know how
to use `x`, or to do standalone assignments. The goal of this commit is
to establish a framework for expression parsing, and to get the vm to
understand how to do assignments.

The interpreter is a mess right now. I'm holding off on testing and
tidying it, because I want to get a better sense of Rust and how I want
this to work before cleaning, optimizing, and testing.