Per-function string and number interning.
I'm not entirely sure I want global symbol interning: Per-function
interning might make it easier to make a repl that doesn't retain unique
strings. Though that depends on how the repl is implemented.

This is a step in the right direction and, more importantly, the
compiler gave me a thumbs up and I am still scared of Rust.